add get window titles function, fix rc returning
