+++ render_service_scripts/launch_max.py
@@ -19,28 +19,6 @@ logger = logging.getLogger(__name__)
 OUTPUT_DIR = 'Output'
 
 
-def get_windows_titles():
-	EnumWindows = ctypes.windll.user32.EnumWindows
-	EnumWindowsProc = ctypes.WINFUNCTYPE(ctypes.c_bool, ctypes.POINTER(ctypes.c_int), ctypes.POINTER(ctypes.c_int))
-	GetWindowText = ctypes.windll.user32.GetWindowTextW
-	GetWindowTextLength = ctypes.windll.user32.GetWindowTextLengthW
-	IsWindowVisible = ctypes.windll.user32.IsWindowVisible
-
-	titles = []
-
-	def foreach_window(hwnd, lParam):
-		if IsWindowVisible(hwnd):
-			length = GetWindowTextLength(hwnd)
-			buff = ctypes.create_unicode_buffer(length + 1)
-			GetWindowText(hwnd, buff, length + 1)
-			titles.append(buff.value)
-		return True
-
-	EnumWindows(EnumWindowsProc(foreach_window), 0)
-
-	return titles
-
-
 def main():
 	args = Util.get_render_args()
 
@@ -97,7 +75,7 @@ def main():
 			fatal_errors_titles = ['Radeon ProRender', 'AMD Radeon ProRender debug assert', os.getcwd() + ' - MAXScript',\
 			'3ds Max', 'Microsoft Visual C++ Runtime Library', \
 			'3ds Max Error Report', '3ds Max application', 'Radeon ProRender Error', 'Image I/O Error', 'Warning', 'Error']
-			error_window = set(fatal_errors_titles).intersection(get_windows_titles())
+			error_window = set(fatal_errors_titles).intersection(util.get_windows_titles())
 			if error_window:
 				rc = -1
 				for child in reversed(p.children(recursive=True)):
@@ -120,7 +98,7 @@ def main():
 
 	# send result data
 	files = util.create_files_dict(OUTPUT_DIR)
-	post_data = util.create_result_status_post_data(rc, OUTPUT_DIR)
+	rc, post_data = util.create_result_status_post_data(rc, OUTPUT_DIR)
 	util.send_status(post_data, files)
 
 	return rc
