+++ render_service_scripts/launch_maya.py
@@ -23,26 +23,7 @@ logger = logging.getLogger(__name__)
 OUTPUT_DIR = 'Output'
 
 
-def get_windows_titles():
-	EnumWindows = ctypes.windll.user32.EnumWindows
-	EnumWindowsProc = ctypes.WINFUNCTYPE(ctypes.c_bool, ctypes.POINTER(ctypes.c_int), ctypes.POINTER(ctypes.c_int))
-	GetWindowText = ctypes.windll.user32.GetWindowTextW
-	GetWindowTextLength = ctypes.windll.user32.GetWindowTextLengthW
-	IsWindowVisible = ctypes.windll.user32.IsWindowVisible
 
-	titles = []
-
-	def foreach_window(hwnd, lParam):
-		if IsWindowVisible(hwnd):
-			length = GetWindowTextLength(hwnd)
-			buff = ctypes.create_unicode_buffer(length + 1)
-			GetWindowText(hwnd, buff, length + 1)
-			titles.append(buff.value)
-		return True
-
-	EnumWindows(EnumWindowsProc(foreach_window), 0)
-
-	return titles
 
 
 def start_monitor_render_thread(args, util):
@@ -202,7 +183,7 @@ def main():
 
 	# send result data
 	files = util.create_files_dict(OUTPUT_DIR)
-	post_data = util.create_result_status_post_data(rc, OUTPUT_DIR)
+	rc, post_data = util.create_result_status_post_data(rc, OUTPUT_DIR)
 	util.send_status(post_data, files)
 
 	return rc
