+++ render_service_scripts/launch_maya_redshift.py
@@ -122,7 +122,7 @@ def main():
 	files = util.create_files_dict(OUTPUT_DIR)
 
 	# send result data
-	post_data = util.create_result_status_post_data(rc, OUTPUT_DIR)
+	rc, post_data = util.create_result_status_post_data(rc, OUTPUT_DIR)
 	util.send_status(post_data, files)
 
 	return rc
